confirm the stack layout
A funtion's stack range is solid.
And every variable is deternmined when the function is executed.
And stack_top and stack_bottom register determined the function
stack range.And call instruction will use the memory after the
stack bottom and use the old funciton's stack_bottom as new
call's stack_top.